apps/social-poster-app/requirements.txt
tweepy
python-dotenv
aiofiles